### 2023-05-21 #5485 @editorialbot
[PRE REVIEW]: EcoAssist: A no-code platform to train and deploy custom YOLOv5 object detection models
**Submitting author:** @PetervanLunteren (<a href="http://orcid.org/0000-0001-5488-4225">Peter van Lunteren</a>)
**Repository:** https://github.com/PetervanLunteren/EcoAssist
**Branch with paper.md** (empty if default branch): JOSS-submission
**Version:** v4.0
