### 2023-05-26 #5485 @PetervanLunteren
@arfon Thanks for letting me know. Perhaps I can help you with finding some suitable reviewers? I had a look at the [JOSS Reviewers Management System](https://reviewers.joss.theoj.org/) and found the following reviewers suitable:

* animikhaich (Animikh Aich)
* lee-seon-woo
* mikemahoney218
* dionhaefner
* tpoisot
* abhilash12iec002
* lgloege
* patrickcgray

Let me know if I can do anything else to assist you.
