### 2023-06-19 #5485 @mstimberg
:wave: @dionhaefner, @smith42, @animikhaich would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
