### 2023-06-19 #5485 @mstimberg
@PetervanLunteren Looking forward to working with you on this submission. While we are waiting for the potential reviewers to respond, you might already have a look at the suggestions made in [this comment](https://github.com/openjournals/joss-reviews/issues/5485#issuecomment-1556274197) and add DOIs in the `.bib` file where appropriate.
