+++ paper.bib
@@ -39,7 +39,8 @@
                experience design; human-computer interaction; image inspection;
                tagging; wildlife monitoring",
   copyright = "http://creativecommons.org/licenses/by/4.0/",
-  language  = "en"
+  language  = "en",
+  doi       = "10.1002/ece3.5767"
 }
 
 @misc{heartexlabs,
@@ -76,7 +77,8 @@
   keywords  = "conservation; ecosystem services; endangered species; population
                extinctions; sixth mass extinction",
   copyright = "https://creativecommons.org/licenses/by/4.0/",
-  language  = "en"
+  language  = "en",
+  doi       = "10.1073/pnas.1922686117"
 }
 
 @article{tuia,
@@ -108,7 +110,8 @@
   month           =  may,
   year            =  2018,
   conference      = "2018 15th Conference on Computer and Robot Vision (CRV)",
-  location        = "Toronto, ON, Canada"
+  location        = "Toronto, ON, Canada",
+  doi             = "10.1109/crv.2018.00052"
 }
 
 @article{gomez,
@@ -121,7 +124,8 @@
   copyright     = "http://arxiv.org/licenses/nonexclusive-distrib/1.0/",
   archivePrefix = "arXiv",
   primaryClass  = "cs.CV",
-  eprint        = "1603.06169"
+  eprint        = "1603.06169",
+  doi           = "10.1016/j.ecoinf.2017.07.004"
 }
 
 @article{norouzzadeh,
