### 2023-06-22 #5581 @mstimberg
👋🏼 @PetervanLunteren, @animikhaich, @oparisot, this is the review thread for the paper. All of our communications will happen here from now on.

There are additional guidelines in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/5581#issue-1769178600).

Please don't hesitate to ping me (@mstimberg) if you have any questions/concerns.
