+++ FAQ.md
@@ -1,0 +1,85 @@
+# Frequently asked questions
+## FAQ 1 - What kind of computer should I use for object detection?
+EcoAssist is available for Windows, Mac and Linux systems. Please note that machine learning can ask quite a lot from your computer in terms of processing power. Besides a minimum of 8GB of RAM, there are no hard system requirements since it is largely hardware-agnostic. However, I would recommend at least 16GB of RAM, but preferably 32GB. Although it will run on an old laptop only designed for text editing, it's probably not going to train any accurate models. Faster machines will analyze images quicker and produce more accurate results. The best models are trained on computers with GPU acceleration. EcoAssist will automatically run on NVIDIA and Apple Silicon GPU if you have the appropriate hardware available, see this GitHub thread for more info. If you have multiple computers at hand, choose the fastest one. Please note that - at the time of writing (May 2023) - training on Apple Silicon GPU is still in beta version.
+
+## FAQ 2 - What should a good training dataset look like?
+It's important to have a well labelled and representative dataset. In computer science there is a concept called GIGO - "garbage in equals garbage out". It means that a model trained on poor quality data will produce poor results. The model can only be as good as the data it learns from. Below are some guidelines for a good dataset provided by the guys at Ultralytics [4].
+* Images per class - 1500 or more images per class recommended.
+* Instances per class - 10000 or more instances (labeled objects) per class recommended.
+* Image variety - Must be representative of deployed environment. For real-world use cases we recommend images from different times of day, different seasons, different weather, different lighting, different angles, different sources (scraped online, collected locally, different cameras) etc.
+* Label consistency - All instances of all classes in all images must be labelled. Partial labelling will not work.
+* Label accuracy - Labels must closely enclose each object. No space should exist between an object and its bounding box. No objects should be missing a label.
+* Background images - Background images are images with no objects that are added to a dataset to reduce false positives. We recommend about 0–10% background images. No labels are required for background images.
+* Class balance - Try to balance the classes in terms of number of instances as much as possible. See labels.jpg after training starts for the class instance bar graph (top left).
+Please note that these are not hard requirements. Don't worry if you don't have thousands of images. You can retrain from an existing model and apply data augmentations when working with limited datasets. See FAQ 4 and 10 for more information about that.
+
+## FAQ 3 - How to choose between transfer learning or training from scratch?
+In machine learning, it is possible to reuse an existing model as the starting point for a new model. For example, the MegaDetector model is excellent at detecting animals in camera trap images. We can transfer the knowledge of what an animal looks like to a new model so that it doesn't have to be trained again. If your dataset is relatively small (i.e., not tens of thousands of images), it's advised to train your model using transfer learning. Of course, there is also the option to train from scratch, but it is usually not recommended. Only train from scratch if you know what you are doing and have a very large dataset (i.e., around 150.000 images or more).
+
+## FAQ 4 - Which model should I choose when transfer learning?
[…]
