+++ FAQ.md
@@ -1,8 +1,8 @@
 # Frequently asked questions
-On this page I answer some some frequently asked questions regarding training object detection models using the YOLOv5 architecture. It corresponds to a medium post on Towards AI.
+On this page I answer some frequently asked questions regarding training object detection models using the YOLOv5 architecture. It corresponds with this medium post on Towards AI (TODO: add link to post after publication).
 
 ## FAQ 1 - What kind of computer should I use for object detection?
-EcoAssist is available for Windows, Mac and Linux systems. Please note that machine learning can ask quite a lot from your computer in terms of processing power. Besides a minimum of 8GB of RAM, there are no hard system requirements since it is largely hardware-agnostic. However, I would recommend at least 16GB of RAM, but preferably 32GB. Although it will run on an old laptop only designed for text editing, it's probably not going to train any accurate models. Faster machines will analyze images quicker and produce more accurate results. The best models are trained on computers with GPU acceleration. EcoAssist will automatically run on NVIDIA and Apple Silicon GPU if you have the appropriate hardware available, see [this thread](https://github.com/PetervanLunteren/EcoAssist#gpu-support) for more info. If you have multiple computers at hand, choose the fastest one. Please note that - at the time of writing (May 2023) - training on Apple Silicon GPU is still in beta version.
+[EcoAssist](https://github.com/PetervanLunteren/EcoAssist) is available for Windows, Mac and Linux systems. Please note that machine learning can ask quite a lot from your computer in terms of processing power. Besides a minimum of 8GB of RAM, there are no hard system requirements since it is largely hardware-agnostic. However, I would recommend at least 16GB of RAM, but preferably 32GB. Although it will run on an old laptop only designed for text editing, it's probably not going to train any accurate models. Faster machines will analyze images quicker and produce more accurate results. The best models are trained on computers with GPU acceleration. EcoAssist will automatically run on NVIDIA and Apple Silicon GPU if you have the appropriate hardware available, see [this thread](https://github.com/PetervanLunteren/EcoAssist#gpu-support) for more info. If you have multiple computers at hand, choose the fastest one. Please note that - at the time of writing (May 2023) - training on Apple Silicon GPU is still in beta version.
 
 ## FAQ 2 - What should a good training dataset look like?
 It's important to have a well labelled and representative dataset. In computer science there is a concept called GIGO - "garbage in equals garbage out". It means that a model trained on poor quality data will produce poor results. The model can only be as good as the data it learns from. Below are some guidelines for a good dataset provided by the guys at [Ultralytics](https://ultralytics.com/).
@@ -47,7 +47,7 @@ It's not uncommon for computers to run out of memory when training. These error 
 ## FAQ 6 - What should I do when I run out of memory?
 When your computer runs out of memory, first close all applications and reboot. Then we'll have to adjust the training parameters to avoid running out of memory again. There are a couple of options to choose from.
 * **Increase paging siz**e - If you are on a Windows computer, setting the paging file (virtual memory) to a fixed size can overcome memory issues. See the steps [here](https://mcci.com/support/guides/how-to-change-the-windows-pagefile-size/) for Windows 10 and [here](https://www.windowscentral.com/software-apps/windows-11/how-to-manage-virtual-memory-on-windows-11) for Windows 11. For the best results and if free storage permits, set the initial size to 1.5 times your RAM capacity and the maximum size to 3 times your RAM capacity. For example, if you have 16 GB of RAM, the initial and maximum sizes should be set to 24.000 MB and 48.000 MB, respectively. Linux and macOS users don't have to worry about this since their memory allocation is arranged differently.
[…]
