+++ EcoAssist_GUI.py
@@ -1,6 +1,6 @@
 # Non-code GUI platform for training and deploying object detection models: https://github.com/PetervanLunteren/EcoAssist
 # Written by Peter van Lunteren
-# Latest edit by Peter van Lunteren on 20 Jun 2023
+# Latest edit by Peter van Lunteren on 28 Jun 2023
 
 # import packages like a christmas tree
 import os
@@ -71,7 +71,7 @@ of_txt = ["of", "de"]
 ##########################################
 
 # post-process files
-def postprocess(src_dir, dst_dir, thresh, sep, file_placement, sep_conf, vis, crp, yol, csv, data_type):
+def postprocess(src_dir, dst_dir, thresh, sep, file_placement, sep_conf, vis, crp, yol, csv, uniquify, label_placement, data_type):
     # log
     print(f"EXECUTED: {sys._getframe().f_code.co_name}({locals()})\n")
 
@@ -91,6 +91,8 @@ def postprocess(src_dir, dst_dir, thresh, sep, file_placement, sep_conf, vis, cr
     global cancel_var
     start_time = time.time()
     nloop = 1
+    timestamp = str(datetime.date.today()) + str(datetime.datetime.now().strftime("%H%M%S"))
+    timestamp = timestamp.replace('-', '')
 
     # warn user
     if data_type == "vid":
@@ -99,6 +101,13 @@ def postprocess(src_dir, dst_dir, thresh, sep, file_placement, sep_conf, vis, cr
                                               ["visualizing, cropping or annotating", "visualizando, recortando o anotando"][lang],
                                               ["visualization, cropping, and annotation", "visualización, recorte y anotación"][lang])
             vis, crp, yol = [False] * 3
+
+    # early exit if user specifies file movement twice (i.e., folder separation and creating folder structure with unique filenames)
+    if yol and uniquify and sep:
+        mb.showerror(error_txt[lang], ["It's not possible to separate folders and create unique filenames at the same time. If you want that, run the post-processing"
+                                       " twice.", "No es posible separar las carpetas y crear nombres de archivo únicos al mismo tiempo. Si deseas lograr eso, ejecuta "
+                                       "el procesamiento posterior dos veces."][lang])
+        return
 
     # fetch label map
     label_map = fetch_label_map_from_json(recognition_file)
@@ -286,7 +295,23 @@ def postprocess(src_dir, dst_dir, thresh, sep, file_placement, sep_conf, vis, cr
         # create yolo annotations
         if yol and len(bbox_info) > 0:
             filename, file_extension = os.path.splitext(file)
-            annot_file = os.path.join(dst_dir, filename + ".txt")
+
+            # uniquify
+            if uniquify:
+                # create unique filenames
+                filename_dst = f"{timestamp}-{'-'.join([x for x in file.split(os.sep) if x != ''])}"
+                annot_file = os.path.join(dst_dir, os.path.splitext(filename_dst)[0] + ".txt")
+                
+                # move files
+                src = os.path.join(src_dir, file)
+                dst = os.path.join(dst_dir, filename_dst)     
+                if label_placement == 1: # move
+                    shutil.move(src, dst)
+                elif label_placement == 2: # copy
+                    shutil.copy2(src, dst)
+            else:
+                annot_file = os.path.join(dst_dir, filename + ".txt")
[…]
