### 2023-06-28 9f4e32f
Update label-animals.md
