### 2023-06-28 5a56c2f
Added exif data to .csv results
