+++ EcoAssist_GUI.py
@@ -17,10 +17,12 @@ import signal
 import shutil
 import platform
 import datetime
+import PIL.Image
 import traceback
 import subprocess
 import webbrowser
 import numpy as np
+import PIL.ExifTags
 import pandas as pd
 import tkinter as tk
 from tkinter import *
@@ -146,13 +148,18 @@ def postprocess(src_dir, dst_dir, thresh, sep, file_placement, sep_conf, vis, cr
         # for files
         csv_for_files = os.path.join(dst_dir, "results_files.csv")
         if not os.path.isfile(csv_for_files):
-            df = pd.DataFrame(list(), columns=["absolute_path", "relative_path", "data_type", "n_detections", "max_confidence"])
+            df = pd.DataFrame(list(), columns=["absolute_path", "relative_path", "data_type", "n_detections", "max_confidence",
+                                               'datetime', 'datetime_original', 'datetime_digitized', 'make', 'shutter_speed_value',
+                                               'aperture_value', 'exposure_bias_value', 'max_aperture_value', 'GPSInfo'])
             df.to_csv(csv_for_files, encoding='utf-8', index=False)
         
         # for detections
         csv_for_detections = os.path.join(dst_dir, "results_detections.csv")
         if not os.path.isfile(csv_for_detections):
-            df = pd.DataFrame(list(), columns=["absolute_path", "relative_path", "data_type", "label", "confidence", "bbox_left", "bbox_top", "bbox_right", "bbox_bottom", "file_height", "file_width"])
+            df = pd.DataFrame(list(), columns=["absolute_path", "relative_path", "data_type", "label", "confidence", "bbox_left",
+                                               "bbox_top", "bbox_right", "bbox_bottom", "file_height", "file_width", 'datetime',
+                                               'datetime_original', 'datetime_digitized', 'make', 'shutter_speed_value', 'aperture_value',
+                                               'exposure_bias_value', 'max_aperture_value', 'GPSInfo'])
             df.to_csv(csv_for_detections, encoding='utf-8', index=False)
 
     # loop through images
@@ -201,6 +208,28 @@ def postprocess(src_dir, dst_dir, thresh, sep, file_placement, sep_conf, vis, cr
                 im_to_crop_path = os.path.join(src_dir, file)
             else:
                 vid = cv2.VideoCapture(os.path.join(src_dir, file))
+
+            # try to read exif data
+            if csv:
+                try:
+                    img_for_exif = PIL.Image.open(os.path.join(src_dir, file))
+                    exif_data = {
+                        PIL.ExifTags.TAGS[k]: v
+                        for k, v in img_for_exif._getexif().items()
+                        if k in PIL.ExifTags.TAGS
+                    }
+                except:
+                    exif_data = None
+
+                # check if datetime values can be found
+                exif_params = []
+                for param in ['DateTime', 'DateTimeOriginal', 'DateTimeDigitized', 'Make', 'ShutterSpeedValue', 'ApertureValue', 'ExposureBiasValue', 'MaxApertureValue', 'GPSInfo']:
+                    try:
+                        param_value = str(exif_data[param])
+                    except:
+                        param_value = "NA"
+                    exif_params.append(param_value)
+                DateTime, DateTimeOriginal, DateTimeDigitized, Make, ShutterSpeedValue, ApertureValue, ExposureBiasValue, MaxApertureValue, GPSInfo = exif_params
 
         # loop through detections
         for detection in image['detections']:
@@ -258,13 +287,13 @@ def postprocess(src_dir, dst_dir, thresh, sep, file_placement, sep_conf, vis, cr
[…]
