### 2023-07-03 4c99306
Copied Exif data from source image during post processing
Copied Exif data from source image during post processing. May be duplicating some of the csv Exif code?

Also fixed bug in cropping code, which resulted in crash if files had been moved.
