+++ EcoAssist_GUI.py
@@ -1736,9 +1736,8 @@ def set_language(to_lang):
         es_widget.config(highlightbackground="black", relief="sunken")
         lang = 1
 
-    # OUTCOMMENT TUTORIAL
-    # # update tutorial text
-    # lbl_tutorial.config(text=lbl_tutorial_txt[lang])
+    # update tutorial text
+    lbl_tutorial.config(text=lbl_tutorial_txt[lang])
 
     # update tab texts
     tabControl.tab(deploy_tab, text=deploy_tab_text[lang])
@@ -2704,13 +2703,12 @@ es_flag = ImageTk.PhotoImage(es_flag)
 es_widget = tk.Button(root, image=es_flag, bg="white", highlightthickness=1, highlightbackground="white", relief="raised", command=lambda: set_language("es"))
 es_widget.grid(column=0, row=1, sticky='e', pady=(0, 2), padx=(3, 43))
 
-# OUTCOMMENT TUTORIAL
 # link to tutorial
-# lbl_tutorial_txt = ['Click here for a step-by-step tutorial on how to use EcoAssist.',
-#                     'Haga clic aquí para ver un tutorial paso a paso sobre cómo usar EcoAssist (en inglés).']
-# lbl_tutorial = Label(master=root, text=lbl_tutorial_txt[lang], anchor="w", bg="white", cursor= "hand2", fg="darkblue", font=(text_font, 13, "underline"))
-# lbl_tutorial.grid(row=1, sticky='ns', pady=2, padx=3)
-# lbl_tutorial.bind("<Button-1>", lambda e:webbrowser.open_new_tab("https://medium.com/@contact_95326/train-and-deploy-custom-object-detection-models-without-a-single-line-of-code-fb99ce5b930e"))
+lbl_tutorial_txt = ['Click here for a step-by-step tutorial on how to use EcoAssist.',
+                    'Haga clic aquí para ver un tutorial paso a paso sobre cómo usar EcoAssist (en inglés).']
+lbl_tutorial = Label(master=root, text=lbl_tutorial_txt[lang], anchor="w", bg="white", cursor= "hand2", fg="darkblue", font=(text_font, 13, "underline"))
+lbl_tutorial.grid(row=1, sticky='ns', pady=2, padx=3)
+lbl_tutorial.bind("<Button-1>", lambda e:webbrowser.open_new_tab("https://medium.com/towards-artificial-intelligence/train-and-deploy-custom-object-detection-models-without-a-single-line-of-code-a65e58b57b03"))
 
 # deploy tab
 deploy_tab = ttk.Frame(tabControl)
