+++ FAQ.md
@@ -1,5 +1,5 @@
 # Frequently asked questions
-On this page I answer some frequently asked questions regarding training object detection models using the YOLOv5 architecture. It corresponds with this medium post on Towards AI (TODO: add link to post after publication).
+On this page I answer some frequently asked questions regarding training object detection models using the YOLOv5 architecture. It corresponds with [this Medium post on Towards AI](https://medium.com/towards-artificial-intelligence/train-and-deploy-custom-object-detection-models-without-a-single-line-of-code-a65e58b57b03).
 
 ## FAQ 1 - What kind of computer should I use for object detection?
 [EcoAssist](https://github.com/PetervanLunteren/EcoAssist) is available for Windows, Mac and Linux systems. Please note that machine learning can ask quite a lot from your computer in terms of processing power. Besides a minimum of 8GB of RAM, there are no hard system requirements since it is largely hardware-agnostic. However, I would recommend at least 16GB of RAM, but preferably 32GB. Although it will run on an old laptop only designed for text editing, it's probably not going to train any accurate models. Faster machines will analyze images quicker and produce more accurate results. The best models are trained on computers with GPU acceleration. EcoAssist will automatically run on NVIDIA and Apple Silicon GPU if you have the appropriate hardware available, see [this thread](https://github.com/PetervanLunteren/EcoAssist#gpu-support) for more info. If you have multiple computers at hand, choose the fastest one. Please note that - at the time of writing (May 2023) - training on Apple Silicon GPU is still in beta version.
