+++ README.md
@@ -47,12 +47,13 @@ I've written this application in my free evenings and would really appreciate it
 </p>
 
 ## Main features
-* Runs on Windows, Mac, and Linux (no admin rights required)
+* Runs on Windows, Mac, and Linux
+* No admin rights required
 * After installation completely offline
 * English :gb: & Español :es:
-* Train your model using the [YOLOv5](https://github.com/ultralytics/yolov5) architecture  
-* Deploy your model on images or videos
-* Annotate your data using the [labelImg](https://github.com/heartexlabs/labelImg) software
+* Train using the [YOLOv5](https://github.com/ultralytics/yolov5) architecture  
+* Deploy on images or videos
+* Annotate using the [labelImg](https://github.com/heartexlabs/labelImg) software
 * GPU acceleration for NVIDIA and Apple Silicon
 * Post-process your data to
   * separate
