### 2023-07-10 #5581 @mstimberg
:wave: @animikhaich, @oparisot, no rush for performing the actual review of course, but just checking in to see whether there are any questions on your side or whether there's anything else I can do to help you with the review? Please be reminded that JOSS reviews can be iterative, i.e. feel free to open issues or comment here as soon as you come across concerns – there is no need to gather everything into a single review text as commonly done for traditional journals. In the same vein, please feel free to tick boxes in your checklist whenever you are sure about them (the two in the beginning – conflict of interest and code of conduct – would be good first candidates :wink: ). Thank you for your time :pray: !
