+++ README.md
@@ -11,35 +11,37 @@ commit](https://img.shields.io/github/last-commit/PetervanLunteren/EcoAssist)
   <img src="https://github.com/PetervanLunteren/EcoAssist-metadata/blob/main/imgs/logo_large.png" width=60% height="auto" />
 </p>
 
-## Introduction
-EcoAssist is an application designed to make life easier for people who want to work with object detection models. Thanks to the people at <a href="https://ultralytics.com/">Ultralytics</a>, it's possible to create models which can locate objects of your interest in images. The only problem is that you need to know a bit of coding before you can use it. That is where EcoAssist comes in handy. It's a no-code interface which makes it easy for everybody. Annotate, train, and deploy your models without a single line of code.
+<h1 align="center">Simplifying camera trap image analysis for ecologists</h1>
 
-[<img align="right" alt="alt_text" width="30%" src="https://github.com/PetervanLunteren/EcoAssist-metadata/blob/main/imgs/Logo_Social_Smartparks-01.png" /> ](https://www.smartparks.org/)
+EcoAssist is an open-source application designed to streamline the work of ecologists dealing with camera trap images. It's an AI platform that enables annotation, training, and deployment of custom models for automatic species detection, offering ecologists a way to save time reviewing images and focus on conservation efforts.
 
-EcoAssist was initially created with the aim of assisting ecological projects (hence the name), but it will handle any kind of object. If you’re not an ecologist and not interested in animals, you’ll just have to input images of your object of interest (blood cells, traffic signs, plant diseases - whatever you want) and ignore the references to the <a href="https://github.com/ecologize/CameraTraps/blob/main/megadetector.md">MegaDetector</a> model. The rest will be exactly the same.
+After a year of development during my free time, I have joined forces with [Smart Parks](https://www.smartparks.org/). We’re working on expanding the software to create a robust toolkit for camera trap image analysis to be used by ecologists worldwide. You can support this initiative by contributing through the sponsorship button below. Please help me to keep improving EcoAssist and let me know about any improvements, bugs, or new features so that I can keep it up-to-date. You can [raise an issue](https://github.com/PetervanLunteren/EcoAssist/issues/new) or [email me](mailto:petervanlunteren@hotmail.com). An e-mail just to say hi and tell me about your project is also very much appreciated!
 
-I would really appreciate it if people would let me know when it's used, and what for. You can contact me at [petervanlunteren@hotmail.com](mailto:petervanlunteren@hotmail.com). Please also help me to keep improving EcoAssist and let me know about any improvements, bugs, or new features so that I can keep it up-to-date.
+[<img align="right" alt="alt_text" width="19%" src="https://github.com/PetervanLunteren/EcoAssist-metadata/blob/main/imgs/Logo_Social_Smartparks-01.png" /> ](https://www.smartparks.org/)
 
-This software is created in collaboration with [Smart Parks](https://www.smartparks.org/).
-
+Our current focus is:
+* Implementing a human-in-the-loop feature for result verification.
+* Improving the annotation process to make it more robust.
+* Testing the setup with a real-world use-case for [the Desert Lion Conservation](https://www.desertlion.info/) project.
+* Set up personalized assistance to support ecologists in effectively using EcoAssist for their projects.
+* Exploring the possibility of providing optimized hardware support.
 
 ## Quick links
 1. [Demo](#demo)
 2. [Overview](#overview)
 3. [Main features](#main-features)
-4. [Extra features for ecologists](#extra-features-for-ecologists)
[…]
