+++ README.md
@@ -17,7 +17,9 @@ EcoAssist is an open-source application designed to streamline the work of ecolo
 
 [<img align="right" alt="alt_text" width="25%" src="https://github.com/PetervanLunteren/EcoAssist-metadata/blob/main/imgs/Logo_Social_Smartparks-01.png" /> ](https://www.smartparks.org/)
 
-After a year of development during my free time, I have joined forces with [Smart Parks](https://www.smartparks.org/). We’re working on expanding the software to create a robust toolkit for camera trap image analysis to be used by ecologists worldwide. You can support this initiative by contributing through the sponsorship button below. Our current focus is:
+After a year of development during my free time, I have joined forces with [Smart Parks](https://www.smartparks.org/). We’re working on expanding the software to create a robust toolkit for camera trap image analysis to be used by ecologists worldwide.
+<!--- You can support this initiative by contributing through the sponsorship button below. --->
+Our current focus is:
 * Implementing a human-in-the-loop feature for result verification.
 * Improving the annotation process to make it more robust.
 * Testing the setup with a real-world use-case for [the Desert Lion Conservation](https://www.desertlion.info/) project.
