### 2023-07-26 2b127b2
Update errors.md
