### 2023-07-27 #5581 @mstimberg
:wave: @oparisot did you have the time to look at the EcoAssist software/paper yet?
