### 2023-07-30 #5581 @PetervanLunteren
@mstimberg @animikhaich @oparisot Thanks for your time! I'm not too familiar with the review process. Are the unchecked boxes areas where the software performs poorly? In other words, are the unchecked boxes actions for me to revise?
