### 2023-07-30 #5581 @mstimberg
@PetervanLunteren Unchecked boxes _could_ mean that there is something to improve in the software/paper, but in that case the reviewers would either leave more specific comments here or open issues in your repository. In this case, it most likely simply means that the reviewers did not yet have the time to verify the unchecked criteria in detail.
