### 2023-07-30 #5581 @oparisot
Hello @PetervanLunteren !!

I checked again the list -- I forgot to check two points, my review is up to date now.
My general feedback: a great work, congrats for this software!

Two minor points:
- Maybe I'm wrong, but I dont' see automated tests in the git repo. Can you confirm?
- Are there similar software? If yes, what is the difference with EcoAssist?

Olivier
