### 2023-07-30 #5581 @animikhaich
Hi @PetervanLunteren,

My checklist is updated. There is an error that is preventing me from running the standard test case and moving forward. I've opened an issue [here](https://github.com/PetervanLunteren/EcoAssist/issues/16). Please address the same and let me know if you need anything else.

A couple of points that need to be addressed:

- The repository is missing a `CONTRIBUTING.md` file, which is required for the "Community Guidelines" checklist. You can refer to [PyTorch's CONTRIBUTING.md](https://github.com/pytorch/pytorch/blob/main/CONTRIBUTING.md) as a reference on how to create one.
- While not necessary, I recommend you containerize the program and add a `Dockerfile` to prevent installation/platform-based limitations like the issue I opened up. It also allows users to install and use the software without installing dependencies on the native system. You can learn more about Docker [here](https://docker-curriculum.com/). Further, you can also host your container on [Docker Hub](https://hub.docker.com/), which will allow users to simply download and run the container instead of having to build the container locally.

Best,
Animikh
