### 2023-07-31 #5581 @PetervanLunteren
Hi @oparisot, Thanks very much for your time! Please see my responses below.

> * Maybe I'm wrong, but I dont' see automated tests in the git repo. Can you confirm?

Indeed, there are currently no automated tests available. I have included [documented manual steps](https://github.com/PetervanLunteren/EcoAssist#test-your-installation) for this purpose. I thought/hoped this to be sufficient for the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#tests), but if not: let me know. In that case I'll revise.

> * Are there similar software? If yes, what is the difference with EcoAssist?

Yes, there are similar packages available. I briefly discuss them in the [paper](https://github.com/openjournals/joss-papers/blob/joss.05581/joss.05581/10.21105.joss.05581.pdf) (line 36-40). Do you advise me to expand on this?

Thanks,
Peter
