### 2023-07-31 #5581 @oparisot
Hello @PetervanLunteren ,

Ok for me for the tests, the documentation is sufficient.
About similar packages, I just suggest to add a short sentence into the README file -- it's a Nice To Have.

Olivier
