### 2023-07-31 a849516
Create CONTRIBUTING.md
