+++ CONTRIBUTING.md
@@ -1,0 +1,100 @@
+# Contributing to EcoAssist
+
+First off, thanks for taking the time to contribute!
+
+All types of contributions are encouraged and valued. See the [Table of Contents](#table-of-contents) for different ways to help and details about how this project handles them. Please make sure to read the relevant section before making your contribution. It will make it a lot easier for us maintainers and smooth out the experience for all involved. The community looks forward to your contributions.
+
+> And if you like the project, but just don't have time to contribute, that's fine. There are other easy ways to support the project and show your appreciation, which we would also be very happy about:
+> - Star the project
+> - Cite the package in your literature
+> - Post about it
+> - Refer this project in your project's readme
+> - Mention the project at local meetups and tell your friends/colleagues
+
+## Table of Contents
+
+- [Code of Conduct](#code-of-conduct)
+- [I Have a Question](#i-have-a-question)
+- [I Want To Contribute](#i-want-to-contribute)
+  - [Reporting Bugs](#reporting-bugs)
+  - [Suggesting Enhancements](#suggesting-enhancements)
+  - [Improving The Documentation](#improving-the-documentation)
+- [Join The Project Team](#join-the-project-team)
+
+## Code of Conduct
+
+This project and everyone participating in it is governed by the
+[EcoAssist Code of Conduct](https://github.com/PetervanLunteren/EcoAssistblob/master/CODE_OF_CONDUCT.md).
+By participating, you are expected to uphold this code. Please report unacceptable behavior
+to <contact@pvanlunteren.com>.
+
+## I Have a Question
+
+> If you want to ask a question, we assume that you have read the [readme](https://github.com/PetervanLunteren/EcoAssist/blob/main/README.md), [error-list](https://github.com/PetervanLunteren/EcoAssist/blob/main/errors.md), and [frequently asked questions](https://github.com/PetervanLunteren/EcoAssist/blob/main/FAQ.md).
+
+Before you ask a question, it is best to search for existing [Issues](https://github.com/PetervanLunteren/EcoAssist/issues) that might help you. In case you have found a suitable issue and still need clarification, you can write your question in that specific issue.
+
+If you then still feel the need to ask a question and need clarification, we recommend the following:
+
+- Open an [Issue](https://github.com/PetervanLunteren/EcoAssist/issues/new).
+- Provide as much context as you can about what you're running into. Which features did you execute, and in which order.
+
+We will then take care of the issue as soon as possible.
+
+## I Want To Contribute
+
+### Reporting Bugs
+
+#### Before Submitting a Bug Report
+
+A good bug report shouldn't leave others needing to chase you up for more information. Therefore, we ask you to investigate carefully, collect information and describe the issue in detail in your report. Please make sure that you are using the latest version before submitting a bug report.
+
+#### How Do I Submit a Good Bug Report?
+
+> You must never report security related issues, vulnerabilities or bugs including sensitive information in public. Instead sensitive bugs must be sent by email to <contact@pvanlunteren.com>.
+
+We use GitHub issues to track bugs and errors. If you run into an issue with the project:
+
+- Open an [Issue](https://github.com/PetervanLunteren/EcoAssist/issues/new).
+- Please provide as much context as possible and describe the *reproduction steps* that someone else can follow to recreate the issue on their own.
+
+Once it's filed:
+
+- The project team will label the issue accordingly.
+- A team member will try to reproduce the issue with your provided steps. If there are no reproduction steps or no obvious way to reproduce the issue, the team will ask you for those steps.
+- If the team is able to reproduce the issue, the issue will be left to be implemented by a contributor.
+
+### Suggesting Enhancements
+
[…]
