### 2023-07-31 eb3e609
Create CODE_OF_CONDUCT.md
