### 2023-07-31 #5581 @PetervanLunteren
>  About similar packages, I just suggest to add a short sentence into the README file -- it's a Nice To Have.

Thanks for the tip, I've added a section to the documentation: https://github.com/PetervanLunteren/EcoAssist#similar-software

Peter
