### 2023-07-31 #5581 @PetervanLunteren
Hi @animikhaich, thanks for your time and suggestions! I'll respond to them below.

> There is an error that is preventing me from running the standard test case and moving forward. I've opened an issue [here](https://github.com/PetervanLunteren/EcoAssist/issues/16). Please address the same and let me know if you need anything else.

Thanks for opening this issue and providing the detailed information. I've looked into it and it seems that it is not directly an EcoAssist issue, but rather a MegaDetector issue regarding the CUDA software that comes with the MegaDetector installation. I've opened an issue at the relevant repository and hope to resolve your problem together with the developers there. Any further communication about this will happen in the issue https://github.com/PetervanLunteren/EcoAssist/issues/16.

> * The repository is missing a `CONTRIBUTING.md` file, which is required for the "Community Guidelines" checklist. You can refer to [PyTorch's CONTRIBUTING.md](https://github.com/pytorch/pytorch/blob/main/CONTRIBUTING.md) as a reference on how to create one.

Thanks for letting me know. Please see the updated repository, where I've updated [the documentation](https://github.com/PetervanLunteren/EcoAssist#contributors) and added a [`CONTRIBUTING.md`](https://github.com/PetervanLunteren/EcoAssist/blob/main/CONTRIBUTING.md) and [`CODE_OF_CONDUCT.md`](https://github.com/PetervanLunteren/EcoAssist/blob/main/CODE_OF_CONDUCT.md).

> * While not necessary, I recommend you containerize the program and add a `Dockerfile` to prevent installation/platform-based limitations like the issue I opened up. It also allows users to install and use the software without installing dependencies on the native system. You can learn more about Docker [here](https://docker-curriculum.com/). Further, you can also host your container on [Docker Hub](https://hub.docker.com/), which will allow users to simply download and run the container instead of having to build the container locally.

This does indeed look great! I will look into this to make the installation more robust. However, since containerizing the program does ask for quite some time investment and the moment I do not have much spare time, I can't promise an ETA. Since it is not necessary for the JOSS publication, I'll add it to my TODO list and will look into it when time permits. I really appreciate the tip!
