### 2023-07-31 #5581 @animikhaich
@PetervanLunteren, Thanks for addressing the issues. I've updated the checklist based on the CONTRIBUTING guidelines.

Since I cannot test the code on my GPU system due to the CUDA Error, I'll test it out on my Macbook and update the checklist based on the same.
