### 2023-07-31 PetervanLunteren/EcoAssist#16 @agentmorris
Hmmm... I haven't seen this before.  I'm about 80% sure this is not really a dimensionality issue, it's a CUDA version mismatch, where for some reason the system CUDA environment is being used instead of the Python environment's CUDA environment.  A couple things to try, starting with the easiest:

1. In the shell from which you are launching EcoAssist, try running:<br/><br/>export LD_LIBRARY_PATH=''<br/><br/>...prior to starting EcoAssist.  I'm 61% sure this will fix the problem, and if that's the case, we have an easy fix, and I get to grumble about how I wish CUDA installs wouldn't mess with LD_LIBRARY_PATH.

2. It would help debug a little if we could take EcoAssist out of the loop just to remove a level of indirection, so if the person who owns the environment is up for it, it would be great to go through the [MegaDetector setup instructions](https://github.com/agentmorris/MegaDetector/blob/main/megadetector.md#using-the-model).  If we can repro the issue there,  we'll have a simpler time debugging.

3. I don't really recommend that the environment owner do this, but FWIW, I think uninstalling CUDA entirely from the system will fix the issue.  In principle I'd like to do this as a debugging step, but it's a big hammer to wield if the user is using the system CUDA for other things.

4. I don't think we'll go past (2) just yet, but if (1) doesn't work, and we can repro the problem in a standalone Python environment (i.e., outside of EcoAssist), we can try to upgrade PyTorch in that environment to match the system CUDA version.  If that works, we've at least verified that it was really a CUDA version mismatch, then we can decide what to do about it.
