### 2023-07-31 PetervanLunteren/EcoAssist#16 @PetervanLunteren
@agentmorris Thanks for your response!

With regards to option 1, the easiest way to run `export LD_LIBRARY_PATH=''` prior to opening EcoAssist would be to add this line somewhere before the [python command on line 109](https://github.com/PetervanLunteren/EcoAssist/blob/main/open.command#L109) in `/home/ani/.EcoAssist_files/EcoAssist/open.command`.
