### 2023-07-31 bac8597
set environment variable `LD_LIBRARY_PATH`
Set environment variable to ensure that the Python environment's CUDA environment is being used instead of the system CUDA environment. Resolves https://github.com/PetervanLunteren/EcoAssist/issues/16 and https://github.com/agentmorris/MegaDetector/issues/106
