### 2023-07-31 #5581 @PetervanLunteren
@mstimberg  for traceability, issue https://github.com/PetervanLunteren/EcoAssist/issues/16 opened by @animikhaich is now resolved with commit https://github.com/PetervanLunteren/EcoAssist/commit/bac8597ecf8623b339467b71864a27d34462e13e

@animikhaich @oparisot thank you very much for your valuable insights!
