+++ paper.bib
@@ -98,7 +98,8 @@
   month     =  feb,
   year      =  2022,
   copyright = "https://creativecommons.org/licenses/by/4.0",
-  language  = "en"
+  language  = "en",
+  doi       = "10.1038/s41467-022-27980-y"
 }
 
 @inproceedings{schneider,
@@ -118,13 +119,12 @@
   title         = "Towards automatic wild animal monitoring: Identification of
                    animal species in camera-trap images using very deep
                    convolutional neural networks",
-  author        = "Gomez, Alexander and Salazar, Augusto and Vargas, Francisco",
+  author        = "Alexander {Gomez Villa} and Augusto Salazar and Francisco Vargas",
   month         =  mar,
-  year          =  2016,
-  copyright     = "http://arxiv.org/licenses/nonexclusive-distrib/1.0/",
-  archivePrefix = "arXiv",
-  primaryClass  = "cs.CV",
-  eprint        = "1603.06169",
+  journal       = "Ecological Informatics",
+  volume        = 41,
+  pages         = "24-32",
+  year          = 2017,
   doi           = "10.1016/j.ecoinf.2017.07.004"
 }
 
