### 2023-08-01 #5581 @mstimberg
@PetervanLunteren All looks very good to me, I only open a small PR fixing two of the references.

Just to check: am I correct that the software itself did not change during the review process, i.e. version v4.0 and the Zenodo archive at https://doi.org/10.5281/zenodo.7223363 are still the most recent ones?
