### 2023-08-01 #5581 @mstimberg
Could you please manually fix some of the metadata in the Zenodo archive (no need to to a new tag/release)? In general, we prefer to have the archive metadata to be as consistent as possible with the JOSS paper. In particular, could you make the titles and author lists match? If you want to leave `@ehallein` in the author list on the Zenodo archive for completeness, that would be fine as well. Also, could you please:
- add your ORCID on Zenodo
- chose the correct license

If you want to make sure that this information will be correctly filled out in the feature when using the automatic GitHub-Zenodo bridge, you can specify them in a `.zenodo.json` file (https://developers.zenodo.org/?python#github).
