### 2023-08-01 #5581 @mstimberg
Excellent work everyone, thanks again @PetervanLunteren for choosing to publish with JOSS, and @animikhaich and @oparisot for taking the time to review the software. I will now recommend acceptance and trigger the final steps in the publication process (after confirmation by an editor-in-chief) :tada: .
