### 2024-12-04 #5581 @PetervanLunteren
Hi Marcel Stimberg, @mstimberg

Due to a trade name conflict, I have to change the name of the EcoAssist software package. Is it possible to update the current paper with the new name and repo? Or do you advise to submit a new one?

Thanks!

Kind regards,

Peter
