### 2024-12-10 #5581 @mstimberg
@PetervanLunteren, I checked back with other editors. JOSS does not have a general update mechanism for papers, but it can update them under exceptional circumstances (retractions, major corrections, …). I wouldn't say that a simple rename of a package would be sufficient, but if e.g. the current name infringes upon a trademark and there are legal reasons to remove/rename the paper, this would be another story. If you think this is the case, please contact me over email with more details, and I will forward this information to the editor-in-chief. Regarding a resubmission: in the same vein, a simple rename would not be enough to warrant a resubmission. You *can* resubmit a new version of your software if it has substantially improved and has major new functionality not covered in the original paper.

My general suggestions for handling a changed name would be:
1. Use [Github's rename functionality](https://docs.github.com/en/repositories/creating-and-managing-repositories/renaming-a-repository) to rename the repository (so that JOSS's "Software Repository" link will be forwarded to the new location)
2. Add a note to the description of the Zenodo archive, stating that the name changed since publication
3. Whenever you cite your software (or when recommending others how to cite it), use "NEW_NAME, previously known as OLD_NAME  [JOSS reference]", so that there is no confusion around the mismatched name in the title.

Please let me know how you'd like to proceed.
