### 2025-03-06 #5581 @PetervanLunteren
@mstimberg Thanks for assisting me with this name change! As you suggested over email, I recovered the [JOSS-submission branch](https://github.com/PetervanLunteren/AddaxAI/tree/JOSS-submission) and retitled the paper.

You can see all the changes here: https://github.com/PetervanLunteren/AddaxAI/commits/JOSS-submission/. In short, I have:
* Removed all instances of the old name `EcoAssist`, and replced it with the new name `AddaxAI`.
* Added a footnote to the first mention of the name in the summary.
* Added a note to the figure caption as it is still the old logo in the screenshots.
* The rest is exactly the same as when the paper got accepted.

Although the PDF generation via [the repo's own GitHub actions](https://github.com/PetervanLunteren/AddaxAI/actions/runs/13694120837) runs successfully, the `generate pdf` command for `editorialbot` fails. Any idea what is going on here? It might have to do with the fact that the repository URL has changed from
`https://github.com/PetervanLunteren/EcoAssist/` to `https://github.com/PetervanLunteren/AddaxAI/`. What would you advise in this case?
